Add specular lighting support
Description:
The camera reference used for drawing in the ScenePanel was retrieved and assigned to a static camera position variable in the Material class before calling the function that draws scene objects. This camera position variable is passed to the shader as a uniform named camPos inside the apply function. Lighting calculations, including specular lighting, are done in the shader based on the camera's position.